I don't know what I updated

+++ MyGraphics/Source/SP2.h
@@ -207,10 +207,10 @@ private:
 	bool modeVillain;
 
 	//Customer
-	int startingAmount;
-	int amountSpent;
-	int remaindingAmount;
-	int amountOvershot;
+	float startingAmount;
+	float amountSpent;
+	float remaindingAmount;
+	float amountOvershot;
 	//Villain
 	int objectsDestroyed;
 	bool isCaught;
